feat: physical card pages are now visually okay
