+++ components/FormBody.vue
@@ -1,0 +1,15 @@
+<template>
+  <div class="id-input">
+    <slot></slot>
+  </div>
+</template>
+
+<style>
+.id-input {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  padding: 0px;
+  gap: 16px;
+}
+</style>
