+++ pages/physical_card/[climberId].vue
@@ -39,29 +39,49 @@ async function insertSerialCode() {
 </script>
 
 <template>
-  <div>
-    <div v-if="loggedIn">
-      <p v-if="insertStatus == 'success'">Edukalt salvestatud</p>
-      <form v-else @submit.prevent="insertSerialCode">
-        <label>
-          Kaarti seerianumber
-          <input v-model.trim="cardSerialCode" />
-        </label>
-        <button>Sisesta</button>
-        <p v-if="insertStatus == 'error'">Sisestamise viga!</p>
-      </form>
-    </div>
-    <div v-else>
-      Not logged in
-      <form @submit.prevent="login">
-        <input v-model="credentials.email" type="email" placeholder="Email" />
-        <input
-          v-model="credentials.password"
-          type="password"
-          placeholder="Parool"
-        />
-        <button>Logi sisse</button>
-      </form>
-    </div>
+  <div style="height: 100%; width: 100%; display: flex">
+    <Layout :show-results="false">
+      <template #form>
+        <div v-if="loggedIn">
+          <p v-if="insertStatus == 'success'">Edukalt salvestatud</p>
+          <form v-else @submit.prevent="insertSerialCode">
+            <p>Sisesta kaardi seerianumber</p>
+            <form-body>
+              <label>
+                Kaardi seerianumber
+                <input v-model.trim="cardSerialCode" />
+              </label>
+              <button>Sisesta</button>
+              <p v-if="insertStatus == 'error'">Sisestamise viga!</p>
+            </form-body>
+          </form>
+        </div>
+        <div v-else>
+          <p>Logi sisse</p>
+          <form @submit.prevent="login">
+            <form-body>
+              <label>
+                Email
+                <input
+                  v-model="credentials.email"
+                  type="email"
+                  placeholder="admin@admin.com"
+                />
+              </label>
+              <label>
+                Parool
+                <input
+                  v-model="credentials.password"
+                  type="password"
+                  placeholder="iamtheadmin"
+                />
+              </label>
+
+              <button>Logi sisse</button>
+            </form-body>
+          </form>
+        </div>
+      </template>
+    </Layout>
   </div>
 </template>
